[FIX] Detach AcessPatterns with unknown filenames
After RAID failure, LVM filenames (dm*, mapper/) may stay alive in the
system. So AccessPattern is binded to really existent filename. However,
if it is unknown (we can not determine what LV/VG it is) then forcefully
unbind it with SCST daemon configuration applying. SCST has to leave
this filename free.
Signed-off-by: Sergey Matveev <stargrave@stargrave.org>